% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_param.R
\name{CV_param}
\alias{CV_param}
\title{Cross validation to choose lambda}
\usage{
CV_param(dataset = data, dat_list = dat_list, penalty = penalty, k = k, n = n)
}
\arguments{
\item{dataset}{\code{data.frame} The dataset you want to split into k-folds and the label column of your dataset should be in the last column.}

\item{dat_list}{\code{list} A list contains the line numbers of every fold.}

\item{penalty}{\code{vector} A vector contains a series of lambda to regularize.}

\item{k}{\code{numeric} The k-folds you want to split your training data into.}

\item{n}{\code{numeric} The number of lambda you have.}
}
\value{
cv_out \code{list} Contains the mean of auc of k models trained by each lambda and the lambda that maximizes auc.
}
\description{
Cross validation to choose the best lambda that maximizes auc.
}
\examples{
dat = iris \%>\% dplyr::filter(Species != "setosa")
dat$Species = ifelse(dat$Species == "versicolor",0,1)
penalty = 10^seq(-4, -1, length.out = 10)
cv_dat = CV_split(10,100,2020)
CV_param(dataset = dat,dat_list = cv_dat,penalty = penalty,k = 10 , n = 10)
}
