% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_split.R
\name{CV_split}
\alias{CV_split}
\title{Data split}
\usage{
CV_split(k, datasize, seed)
}
\arguments{
\item{k}{\code{numeric} The k-folds you want to split your training data into.}

\item{datasize}{\code{numeric}  Your training data size.}

\item{seed}{\code{numeric} A seed to sample.}
}
\value{
data_list \code{list} Contains line numbers of k-folds data.
}
\description{
Line numbers of dataset split to k-folds randomly.
}
\examples{
CV_split(10,100,2020)
}
