% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_test.R
\name{final_test}
\alias{final_test}
\title{Train model again and evaluate the performance of the model.}
\usage{
final_test(
  data_train = data_train,
  data_test = data_test,
  lambda = lambda,
  maxIterNum = 1e+05,
  step = 5e-04
)
}
\arguments{
\item{data_train}{\code{data.frame} All training dataset.}

\item{data_test}{\code{data.frame} All test dataset.}

\item{lambda}{\code{numeric} The best lambda that maximizes auc.}

\item{maxIterNum}{\code{numeric} Maximum number of iterations.}

\item{step}{\code{numeric} Gradient descent stepsize.}
}
\value{
final_res \code{list} Contains the regression coefficients,prediction probability and category,confusion matrix,recall and precision.
}
\description{
Train model again in all training dataset and evaluate the performance of the model in the test dataset.
}
\examples{
set.seed(2020)
dat = iris \%>\% dplyr::filter(Species != "setosa")
dat$Species = ifelse(dat$Species == "versicolor",0,1)
sam_id = sample(1:nrow(dat),nrow(dat)*0.75,replace = FALSE)
df_train = dat[sam_id,]
df_test = dat[-sam_id,]
final_test(data_train = df_train,data_test = df_test,
            lambda = 0.1, maxIterNum = 100000,step = 0.0005)
}
